+++ src/constants.ts
@@ -1,3 +1,11 @@
 export const CLI_NAME = 'ztd'
 
 export const DEBUG_PREFIX = 'ztd'
+
+export const ERROR_PREFIX = '❌'
+
+export const INFO_PREFIX = '🛈'
+
+export const LINK_ISSUE_BUG = `https://github.com/jackdbd/zod-to-doc/labels/bug`
+
+export const LINK_ISSUE_ENHANCEMENT = `https://github.com/jackdbd/zod-to-doc/labels/enhancement`